Create githook service from pps master
And properly pretty print InspectPipeline results
